pyproject.toml
[tool.poetry]
name = "fastapi-sqlalchemy-boilerplate"
version = "0.1.0"
description = "FastAPI SQLAlchemy async boilerplate with configured migrations and admin panel"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
sqladmin = "^0.15.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
email-validator = "^2.1.0"
jinja2 = "^3.1.2"
watchfiles = "^0.21.0"
websockets = "^12.0"
wtforms = "^3.1.1"
orjson = "^3.9.10"
ujson = "^5.9.0"
httpx = "^0.25.2"
starlette = "^0.27.0"
anyio = "^3.7.1"
click = "^8.1.7"
dnspython = "^2.4.2"
greenlet = "^3.0.1"
h11 = "^0.14.0"
httpcore = "^1.0.2"
httptools = "^0.6.1"
idna = "^3.6"
itsdangerous = "^2.1.2"
mako = "^1.3.0"
markupsafe = "^2.1.3"
pyyaml = "^6.0.1"
sniffio = "^1.3.0"
typing-extensions = "^4.8.0"
certifi = "^2023.11.17"
pyodbc = "^5.2.0"
sqlmodel = "^0.0.24"
dependency-injector = "^4.48.1"
ruff = "^0.14.9"
passlib = {extras = ["argon2"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
bcrypt = "^5.0.0"


[tool.poetry.group.dev]
optional = true


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
ruff = "^0.14.9"
mypy = "^1.18.2"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
pytest-sugar = "^1.0.0"
pytest-asyncio = "^0.25.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# ---------------------------
# Ruff (lint + imports)
# ---------------------------
[tool.ruff]
line-length = 120
src = ["src", "tests"]
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.lint.isort]
combine-as-imports = true
force-single-line = false
force-sort-within-sections = true
known-first-party = ["src"]
lines-after-imports = 2
split-on-trailing-comma = true


[tool.pytest.ini_options]
pythonpath = ["src"]


[tool.mypy]
python_version = "3.12"
namespace_packages = true
show_error_codes = true
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = [
    "sqlalchemy.*",
    "alembic.*",
    "fastapi.*",
    "uvicorn.*",
    "pydantic.*",
    "pydantic_settings.*",
]
ignore_missing_imports = true
